Provide configData to hawkbit server
:Release Notes:
Provide configData to hawkbit server

:Detailed Notes:
A configData link is included in the polling message, when the device is
registered or the upgrade is complete. So, if the link is given in the
polling message, we just send configData.

:Testing Performed:
Verify the target is created and its information is registerd in the
Attributes tab, after executing the commands below.
$ fw_setenv hawkbit_tenant DEFAULT
$ fw_setenv hawkbit_url http://10.177.247.56:8080
$ fw_setenv hawkbit_token 81ec98d166b3cf6e5b812f51b2bb3c8d
$ LOG_TYPE=console LOG_LEVEL=verbose swupdater

:QA Notes:

:Issues Addressed:
[PLAT-81456] Provide configData to hawkbit server

Change-Id: I6a898bc927737da44471017eef2ad9db2e3a1f40
Reviewed-on: http://gpro.lge.com/247572
Reviewed-by: Myungchul Kim <myungchul.kim@lge.com>
Tested-by: Myungchul Kim <myungchul.kim@lge.com>
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>
Tested-by: Sangwoo Kang <sangwoo82.kang@lge.com>