Implement rollback on boot failure on raspberrypi
:Release Notes:
Implement rollback on boot failure on raspberrypi

:Detailed Notes:
Implement rollback with UBoot's environment variables.
We use the following variables for rollback.
- bootlimit
- bootcount
- rollback
- upgrade_available
UBoot increments the bootcount by 1 for every boot.
And if bootcount exceeds the bootlimit,
UBoot set rollback = 1 and boots with the rollback deployment.
These all work only when upgrade_available is 1.
And once the rollback becomes 1, the bootcount is no longer checked,
and booted directly with the rollback deployment on the next boot.

:Testing Performed:
1. normal case
  - Delta update with hawkbit server
  - reboot
  - Verify the update action finished successfully on the server.
2. rollback case
  - Delta update with hawkbit server
  - Modify boot param to fail to boot.
    sed -i '/^kernel_image=/ s/$/-test/' /boot/loader/uEnv.txt
  - reboot
  - Verify the update action finished with an error on the server.

:QA Notes:

:Issues Addressed:
[PLAT-81751] Implement rollback on boot failure on raspberrypi

Change-Id: Ie4426f015eb41cc6c1a3b37f3ac722c76ec389f1
Reviewed-on: http://gpro.lge.com/248919
Reviewed-by: Myungchul Kim <myungchul.kim@lge.com>
Tested-by: Myungchul Kim <myungchul.kim@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>