Restore to previous install status, if actionHistory exists
:Release Notes:
Restore to previous install status, if actionHistory exists

:Detailed Notes:
When swupdater look up a deployment (install action), if there exists 'actionHistory',
this means that the installation was started before, but not completed.
Swupdater then restores the previous installation status and resumes.
When restoring, specifically 'completed' status calls several callbacks,
So this commit adds an option to not call the callbacks.

:Testing Performed:
Kill swupdater while installing some updates
Verify the install resumes when swupdater is restarted.

:QA Notes:

:Issues Addressed:
[PLAT-83229] Suspend/Resume Installation

Change-Id: I7a134ac9e2e581babe7a919e36c72472f13aa85b
Reviewed-on: http://gpro.lge.com/252880
Reviewed-by: Myungchul Kim <myungchul.kim@lge.com>
Tested-by: Myungchul Kim <myungchul.kim@lge.com>
Reviewed-by: Sangwoo Kang <sangwoo82.kang@lge.com>