Add a script which generates a delta and uploads it to hawkBit.
:Release Notes:
Add a script which generates a delta and uploads it to hawkBit.

:Detailed Notes:
This script depends on ostree, curl and jq (command line json handler).
So, to execute this, you need to install the above.

If installed above, first initialize the local ostree repo.
$ mkdir -p /path_to/ostree_repo_official
$ ./delta.sh --repo=/path_to/ostree_repo_official --prepare
The prepare command only needs to be executed once the first time.

Next, run the script without prepare option, specifying the same repo.
$ ./delta.sh --repo=/path_to/ostree_repo_official --addr=http://10.177.247.56:8888
This command does the followings:
- ostree pull --repo=$Repo official:${Branch} --depth=5
- ostree static-delta generate --repo=$Repo --min-fallback-size=0 --inline
                               --from=$From --to=$To --filename=$Filename
- Create a softwaremodule and upload the delta
- Create a distribution and assign the softwaremodule to it
- Create a filter and set the distribution to auto-assignment.

:Testing Performed:
mkdir -p /path_to/ostree_repo_official
./delta.sh --repo=/path_to/ostree_repo_official --prepare
./delta.sh --repo=/path_to/ostree_repo_official --addr=http://10.177.247.56:8888

:QA Notes:

:Issues Addressed:
[PLAT-86404] Generate delta from ostree repo and upload to hawkBit server

Change-Id: I45ee84eb7bb4b187736eb4c3a8a8cefe2d41fa97