Apply block based incremental upgrade on auto board
:Release Notes:
Apply block based incremental upgrade on auto board

:Detailed Notes:
Block-based incremental upgrade uses a tool called xdelta3.
The 'boot' partition does not use delta and overwrites the entire partition with dd command.
We use delta only for the 'system' partition, and when getting delta,
mount the image with read only and use it as the source for xdelta3. (Refer PLAT-95354)
The delta file should have the extension xd3.

:Testing Performed:
- Generate delta file
  - Mount the image before upgrade as a loop device
    sudo mount -o ro,loop webos-auto-image-1844.ext4 /mnt
  - Check the loop device
    df | grep mnt  # e.g. /dev/loop0
  - Generate delta with xdelta3 with the loop device
    sudo xdelta3 -S none -e -s /dev/loop0 webos-auto-image-1845.ext4 upgrade-1844-1845.xd3
- Upgrade
  - Find current boot slot
    abctl --boot_slot  # _a or _b
  - Find partition by boot slot
    realpath /dev/disk/by-partlabel/system_a  # /dev/sda6 for system_a, /dev/sda7 for system_b
  - Apply xdelta3
    xdelta3 -S none -d -s /dev/sda6 /home/root/upgrade-1844-1845.xd3 /dev/sda7

:QA Notes:

:Issues Addressed:
[PLAT-97444] Apply block based incremental upgrade on auto board

Change-Id: Ie08d4b1e5d81d04a653952cb4af0967da4a51d63